Fixes #3578 -> Logging out of the API should deprecate the token
The request to deprecate the user's token was cancelled by the browser due to the removal of the DOM element that triggered the event.
In this case, the logout button perfomed an inmediate redirect to the '/logout' page, and therefore, removing the element from the DOM, causing the failure of the request.

As a solution, the anchor's behaviour (<a href='logout'..>)(part of X-Pack plugin) has been overrided with event.preventDefault(), to later on navigate programatically to '/logout' once a response from the server has been received.

This commit also applies an automatic formmating to the modified files.